refactor: cleanup postgres extractor (#148)
* refactor: cleanup postgres extractor

* refactor: allow exclude and default exclude dbs
