+++ plugins/extractors/populate.go
@@ -1,8 +1,5 @@
 package extractors
 
 import (
-	_ "github.com/odpf/meteor/plugins/extractors/bigquery"
-	_ "github.com/odpf/meteor/plugins/extractors/csv"
-	_ "github.com/odpf/meteor/plugins/extractors/date"
 	_ "github.com/odpf/meteor/plugins/extractors/postgres"
 )
